requirements.txt
fastapi
uvicorn
torch
open-clip-torch
pillow
requests